dom/html/test/test_bug347174_xsl.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=347174
-->
<head>
  <title>Test for Bug 347174</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=347174">Mozilla Bug 347174</a>
<p id="display"></p>
<div id="content" style="display: none">
  <iframe src="347174transformable.xml" id="iframe"></iframe>
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 347174 **/
// Test of readyState of XML document transformed via XSLT to HTML
// this test passes in IE7
window.readyStateText = [];

function showMessage(msg) {
    window.readyStateText.push(msg);
    document.getElementById("display").innerHTML = readyStateText.join("<br>");
}

function frameScriptTag(readyState) {
  isnot(readyState, "complete", "document.readyState should not be 'complete' when scripts run initially");
  showMessage("script tag: " + readyState);
}

function frameLoad(readyState) {
  is(readyState, "complete", "document.readyState should be 'complete' on load");
  showMessage("load: " + readyState);
  SimpleTest.finish();
}

function frameReadyStateChange(readyState) {
  showMessage("readyStateChange: " + readyState);
}

function frameDOMContentLoaded(readyState) {
  is(readyState, "interactive", "document.readyState should be 'interactive' on DOMContentLoaded");
  showMessage("DOMContentLoaded: " + readyState);
}

SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
